Refactor GTPacketSound to use doubles coordinates instead of ints (#5155)
(cherry picked from commit 21ec49745dbbd62a33798bb2814c0c72f77c9276)

+++ src/main/java/gregtech/api/net/GTPacketInfiniteSpraycan.java
@@ -140,9 +140,9 @@ public class GTPacketInfiniteSpraycan extends GTPacket {
                 SoundResource.GT_SPRAYCAN_SHAKE,
                 1.0F,
                 1,
-                (int) player.posX,
-                (int) player.posY,
-                (int) player.posZ);
+                player.posX,
+                player.posY,
+                player.posZ);
         }
 
         private static void playLockSound(final EntityPlayerMP player) {
@@ -151,9 +151,9 @@ public class GTPacketInfiniteSpraycan extends GTPacket {
                 SoundResource.GT_SPRAYCAN_LOCK,
                 1.0F,
                 1,
-                (int) player.posX,
-                (int) player.posY,
-                (int) player.posZ);
+                player.posX,
+                player.posY,
+                player.posZ);
         }
 
         private static void playUnlockSound(final EntityPlayerMP player) {
@@ -162,9 +162,9 @@ public class GTPacketInfiniteSpraycan extends GTPacket {
                 SoundResource.GT_SPRAYCAN_UNLOCK,
                 1.0F,
                 1,
-                (int) player.posX,
-                (int) player.posY,
-                (int) player.posZ);
+                player.posX,
+                player.posY,
+                player.posZ);
         }
 
         abstract boolean execute(final BehaviourSprayColorInfinite behavior, ItemStack itemStack, EntityPlayerMP player,
